TopSig appears to be working now.  Still need to deal with stop words, stemming, and relevance feedback.
git-svn-id: https://groke.otago.ac.nz/ant/ant@1182 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/index_document_topsig_signature.h
@@ -24,8 +24,8 @@ public:
 	~ANT_index_document_topsig_signature();
 
 	void rewind(void);
-	double add_term(ANT_index_document_topsig *globalstats, ANT_string_pair *term, long long term_frequency, long long document_length, long long collection_length_in_term);
-	double add_term(ANT_index_document_topsig *globalstats, char *term, long long term_frequency, long long document_length, long long collection_length_in_term);
+	unsigned long long add_term(ANT_index_document_topsig *globalstats, ANT_string_pair *term, long long term_frequency, long long document_length, long long collection_length_in_term);
+	unsigned long long add_term(ANT_index_document_topsig *globalstats, char *term, long long term_frequency, long long document_length, long long collection_length_in_term);
 	double *get_vector(void) { return vector; }
 } ;
